feat(app): Created datamodels folder, Moved app_models file
