+++ src/app/agents/agent_system.py
@@ -33,7 +33,7 @@ from pydantic_ai.messages import ModelRequest
 from pydantic_ai.usage import UsageLimits
 
 from app.agents.llm_model_funs import get_api_key, get_models, get_provider_config
-from app.config.data_models import (
+from app.datamodels.app_models import (
     AgentConfig,
     AnalysisResult,
     AppEnv,
