+++ src/app/utils/load_settings.py
@@ -10,7 +10,7 @@ from pathlib import Path
 
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
-from app.config.data_models import ChatConfig
+from app.datamodels.app_models import ChatConfig
 from app.utils.error_messages import (
     failed_to_load_config,
     file_not_found,
